Created new package, nav. Moved MainActivity.kt navigation to nav/AuthNavigation.kt and StartGameActivity.kt navigation to nav/AppNavigation.kt.
