+++ app/src/main/java/com/example/cst438_project_1/StartGameActivity.kt
@@ -5,34 +5,10 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.AlertDialog
-import androidx.compose.material3.Button
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableIntStateOf
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.alpha
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.navigation.compose.NavHost
-import androidx.navigation.compose.composable
-import androidx.navigation.compose.rememberNavController
 import com.example.cst438_project_1.nav.AppNavigation
 import com.example.cst438_project_1.ui.theme.Cst438project1Theme
 
@@ -56,69 +32,5 @@ class StartGameActivity : ComponentActivity() {
         val intent = Intent(this, MainActivity::class.java)
         startActivity(intent)
         finish()
-    }
-}
-
-
-@Composable
-fun GameScreen(modifier: Modifier = Modifier, onQuitClick: () -> Unit = {}) {
-    var score by remember { mutableIntStateOf(0) }
-    var gameOver by remember { mutableStateOf(false) }
-
-
-    val updateScore: () -> Unit = {
-        if(!gameOver) {
-            score++
-        }
-    }
-
-    val wrongAnswer: () -> Unit = {
-        gameOver = true
-    }
-
-    val retryGame: () -> Unit = {
-        gameOver = false
-        score = 0
-    }
-
-    Column(
-        modifier = modifier.fillMaxSize(),
-        verticalArrangement = Arrangement.Center,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        Text("Game Screen Placeholder")
-        Text("Score: $score")
-        Spacer(modifier = Modifier.height(24.dp))
-        Button(
-            onClick = updateScore,
-            enabled = !gameOver
-        ) {
-            Text(text = "Update Score")
-        }
-        Button(
-            onClick = wrongAnswer,
-            enabled = !gameOver
-        ) {
-            Text(text = "Wrong Answer")
-        }
-        Button(
-            onClick = retryGame,
-            modifier = Modifier.alpha(if (gameOver) 1f else 0f),
-            enabled = gameOver
-        ) {
-            Text(text = "Retry Game")
-        }
-
-        Button(onClick = onQuitClick) {
-            Text(text = "Quit")
-        }
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun GameScreenPreview() {
-    Cst438project1Theme {
-        GameScreen()
     }
 }
